# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API DAZZART",
    "description": "Documentación de la API de DAZZART",
    "version": "1.0.0"
  },
  "host": "3.134.246.187",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/productos/listar": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/productos/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/productos/agregar": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "imagen": {
                  "example": "any"
                },
                "numero_serial": {
                  "example": "any"
                },
                "nombre": {
                  "example": "any"
                },
                "descripcion": {
                  "example": "any"
                },
                "precio": {
                  "example": "any"
                },
                "stock": {
                  "example": "any"
                },
                "id_categoria": {
                  "example": "any"
                },
                "id_subcategoria": {
                  "example": "any"
                },
                "fecha_creacion": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/productos/editar/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "numero_serial": {
                  "example": "any"
                },
                "nombre": {
                  "example": "any"
                },
                "descripcion": {
                  "example": "any"
                },
                "precio": {
                  "example": "any"
                },
                "stock": {
                  "example": "any"
                },
                "id_categoria": {
                  "example": "any"
                },
                "id_subcategoria": {
                  "example": "any"
                },
                "imagen": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/productos/eliminar/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/productos/listar-imagenes": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}